Conflicts:
	src/1. Layers/1.3 Domain/Well.Domain/Well.Domain.csproj
	src/1. Layers/1.5 ACL/Well.Adam/App.config
	src/1. Layers/1.5 ACL/Well.Adam/Infrastructure/AdamRouteFileProvider.cs
	src/1. Layers/1.5 ACL/Well.Adam/Well.Adam.csproj
	src/1. Layers/1.5 ACL/Well.TranSend/App.config
	src/4. Test/BDD/App.config
	src/4. Test/BDD/BDD.csproj
	src/4. Test/BDD/Framework/AdamImportConfiguration.cs
	src/4. Test/BDD/Steps/AdamImportSteps.cs

+++ src/3. Database/Well/dbo/Tables/ExceptionAction.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[ExceptionAction]
 (
-	[Id] [TINYINT] IDENTITY(1,1) NOT NULL,
+	[Id] [INT] IDENTITY(1,1) NOT NULL,
 	[Description] VARCHAR(50) NOT NULL,
 	[CreatedBy] VARCHAR(50) NOT NULL,
 	[CreatedDate] DATETIME NOT NULL,
